Updates to organization of test files.
